finish fixing mpl tests
