+++ tests/backends/matplotlib/mpl_interactive_test.py
@@ -3,31 +3,27 @@
 
 import pytest
 
-from plopp import Node
-from plopp.backends.matplotlib.interactive import InteractiveFig
 from plopp.data.testing import data_array
-from plopp.graphics.imageview import ImageView
-from plopp.graphics.lineview import LineView
 
 
 @pytest.mark.usefixtures('_use_ipympl')
 def test_logx_1d_toolbar_button():
     da = data_array(ndim=1)
-    fig = InteractiveFig(LineView, Node(da), scale={'xx': 'log'})
+    fig = da.plot(scale={'xx': 'log'})
     assert fig.toolbar['logx'].value
 
 
 @pytest.mark.usefixtures('_use_ipympl')
 def test_logy_1d_toolbar_button():
     da = data_array(ndim=1)
-    fig = InteractiveFig(LineView, Node(da), norm='log')
+    fig = da.plot(norm='log')
     assert fig.toolbar['logy'].value
 
 
 @pytest.mark.usefixtures('_use_ipympl')
 def test_logxy_2d_toolbar_buttons():
     da = data_array(ndim=2)
-    fig = InteractiveFig(ImageView, Node(da), scale={'xx': 'log', 'yy': 'log'})
+    fig = da.plot(scale={'xx': 'log', 'yy': 'log'})
     assert fig.toolbar['logx'].value
     assert fig.toolbar['logy'].value
 
@@ -35,5 +31,5 @@ def test_logxy_2d_toolbar_buttons():
 @pytest.mark.usefixtures('_use_ipympl')
 def test_log_norm_2d_toolbar_button():
     da = data_array(ndim=2)
-    fig = InteractiveFig(ImageView, Node(da), norm='log')
+    fig = da.plot(norm='log')
     assert fig.toolbar['lognorm'].value
